move to touches sending a floating point note

+++ mecapi/mec_voice.h
@@ -21,7 +21,7 @@ public:
     struct Voice {
         int i_;
         int id_;
-        int note_;
+        float note_;
         float x_;
         float y_;
         float z_;
